fix: resolve implementation for multiple contributors
