+++ lib/parse-comment.js
@@ -1,5 +1,3 @@
-module.exports = parseComment;
-
 const nlp = require("compromise");
 
 // Types that are valid (multi words must all be lower case)
@@ -133,12 +131,11 @@ const plugin = {
 
 nlp.plugin(plugin);
 
-function getMatcher(message, action) {
-  const whoMatched = nlp(message)
-    .match(`${action} [.]`)
-    .normalize({
-      // We cannot use whitespace: true, because that gets rid of trailing hyphens.
-      whitespace: false, // remove hyphens, newlines, and force one space between words
+function findWho(message, action) {
+  function findWhoSafe(match) {
+    message = message.replace("-", "#/#"); // workaround (https://github.com/spencermountain/compromise/issues/726)
+    const whoNormalizeSettings = {
+      whitespace: true, // remove hyphens, newlines, and force one space between words
       case: false, // keep only first-word, and 'entity' titlecasing
       numbers: false, // turn 'seven' to '7'
       punctuation: true, // remove commas, semicolons - but keep sentence-ending punctuation
@@ -150,30 +147,78 @@ function getMatcher(message, action) {
       plurals: false, // turn "batmobiles" into "batmobile"
       verbs: false, // turn all verbs into Infinitive form - "I walked" → "I walk"
       honorifics: false, //turn 'Vice Admiral John Smith' to 'John Smith'
-    })
-    .data()[0]
-    .text.replace(/\s/g, "");
+    };
+    const matchedSet = nlp(message)
+      .match(match)
+      .normalize(whoNormalizeSettings)
+      .data();
 
-  return whoMatched.startsWith("@") ? whoMatched.substr(1) : whoMatched;
+    if (matchedSet.length > 0) {
+      matchedText = matchedSet[0].text;
+      matchedText = matchedText.replace("#/#", "-");
+
+      return matchedText;
+    }
+  }
+
+  const whoMatchedMention = findWhoSafe(`@[.]`);
+  // TODO: Unused
+  // if (whoMatchedMention) {
+  //   return whoMatchedMention;
+  // }
+
+  const whoMatchedByAction = findWhoSafe(`${action} [.]`);
+  if (whoMatchedByAction) {
+    return whoMatchedByAction;
+  }
+
+  const whoMatchedByFor = findWhoSafe(`[.] for`);
+  if (whoMatchedByFor) {
+    return whoMatchedByFor;
+  }
 }
 
-function parseAddComment(message, action) {
-  const who = getMatcher(message, action);
+function parseAddSentence(message, action) {
+  message = message
+    .split(",")
+    .map((e) => e.trim())
+    .join(", ");
+
+  const whoMatched = findWho(message, action);
+  if (!whoMatched) {
+    return {
+      who: undefined,
+    };
+  }
+
+  const who = whoMatched.startsWith("@") ? whoMatched.substr(1) : whoMatched;
 
   // Contributions
   const doc = nlp(message).toLowerCase();
-  // This is to support multi word 'matches' (although the compromise docs say it supports this *confused*)
+  // This is to support multi word 'matches' (altho the compromise docs say it supports this *confused*)
   Object.entries(contributionTypeMultiWordMapping).forEach(
     ([multiWordType, singleWordType]) => {
       doc.replace(multiWordType, singleWordType);
     }
   );
+
+  const rawContributions = doc
+    .match("#Contribution")
+    .data()
+    .map((data) => {
+      // This removes whitespace, commas etc
+      return data.normal;
+    });
+
   const contributions = doc
     .match("#Contribution")
     .data()
     .map((data) => {
       // This removes whitespace, commas etc
       return data.normal;
+    })
+    .filter((element, index) => {
+      return rawContributions.indexOf(element) === index;
     })
     .map((type) => {
       if (contributionTypeMappings[type]) return contributionTypeMappings[type];
@@ -187,9 +232,27 @@ function parseAddComment(message, action) {
     });
 
   return {
-    action: "add",
     who,
     contributions,
+  };
+}
+
+function parseAddComment(message, action) {
+  const contributors = {};
+
+  const sentences = nlp(message).sentences();
+  sentences.forEach(function (sentence) {
+    const sentenceRaw = sentence.data()[0].text;
+    const { who, contributions } = parseAddSentence(sentenceRaw, action);
+
+    if (who) {
+      contributors[who] = contributions;
+    }
+  });
+
+  return {
+    action: "add",
+    contributors,
   };
 }
 
@@ -198,11 +261,13 @@ function parseComment(message) {
 
   const action = doc.toLowerCase().match("#Action").normalize().out("string");
 
-  if (action === "add") {
+  if (action.match("add")) {
     return parseAddComment(message, action);
   }
-
+  
   return {
     action: false,
   };
 }
+
+module.exports = parseComment;
