+++ lib/process-issue-comment.js
@@ -8,9 +8,32 @@ const addContributor = require("./add-contributor");
 
 async function processIssueComment({ context, commentReply }) {
   const commentBody = context.payload.comment.body;
+  const { action, contributors } = parseComment(commentBody);
+
+  if (action !== "add") {
+    // log.info(`Unknown action "${action}"`);
+    commentReply.reply(`I could not determine your intention.`);
+    commentReply.reply(
+      `Basic usage: @all-contributors please add @someone for code, doc and infra`
+    );
+    commentReply.reply(
+      `For other usages see the [documentation](https://allcontributors.org/docs/en/bot/usage)`
+    );
+    return;
+  }
+
+  for (var k in contributors) {
+    who = k
+    contributions = contributors[who]
+
+    await triggerActionAdd(context, commentReply, who, contributions)
+  }
+}
+
+async function triggerActionAdd(context, commentReply, who, contributions) {
+  const action = "add"
   const repo = context.payload.repository;
   const createdBy = context.payload.comment.user;
-  const { who, action, contributions } = parseComment(commentBody);
 
   const log = context.log.child({
     who,
@@ -26,18 +49,6 @@ async function processIssueComment({ context, commentReply }) {
     private: repo.private,
     success: false,
   });
-
-  if (action !== "add") {
-    log.info(`Unknown action "${action}"`);
-    commentReply.reply(`I could not determine your intention.`);
-    commentReply.reply(
-      `Basic usage: @all-contributors please add @someone for code, doc and infra`
-    );
-    commentReply.reply(
-      `For other usages see the [documentation](https://allcontributors.org/docs/en/bot/usage)`
-    );
-    return;
-  }
 
   if (contributions.length === 0) {
     log.info("No contributions");
@@ -58,6 +69,7 @@ async function processIssueComment({ context, commentReply }) {
 
   repository.setSkipCi(config.options.skipCi);
 
+  // TODO: Add implement to support single PR
   const { pullCreated, user } = await addContributor({
     context,
     commentReply,
